Move services into the service directory
