+++ .travis.yml
@@ -10,22 +10,22 @@ jobs:
   include:
     - stage: test
       name: "Test orders service"
-      script: ./gradlew :orders:check
+      script: ./gradlew :service:orders:check
     - name: "Test payments service"
-      script: ./gradlew :payments:check
+      script: ./gradlew :service:payments:check
     - name: "Test stock service"
-      script: ./gradlew :stock:check
+      script: ./gradlew :service:stock:check
     - name: "Test users service"
-      script: ./gradlew :users:check
+      script: ./gradlew :service:users:check
     - stage: build
       name: "Build orders service"
-      script: ./ci/build.sh orders
+      script: ./ci/build.sh service:orders service/orders orders
     - name: "Build payments service"
-      script: ./ci/build.sh payments
+      script: ./ci/build.sh service:payments service/payments payments
     - name: "Build stock service"
-      script: ./ci/build.sh stock
+      script: ./ci/build.sh service:stock service/stock stock
     - name: "Build users service"
-      script: ./ci/build.sh users
+      script: ./ci/build.sh service:users service/users users
     - stage: end-to-end
       name: "Run end-to-end tests"
       script: ./ci/test.sh
