+++ docker-compose.yml
@@ -24,7 +24,7 @@ services:
       - "8080:80"
   orders:
     image: wdmk/orders:${IMAGE_VERSION:-latest}
-    build: orders
+    build: service/orders
     links:
       - kafka
     ports:
@@ -33,7 +33,7 @@ services:
       SPRING_KAFKA_BOOTSTRAP_SERVERS: kafka:9092
   payments:
     image: wdmk/payments:${IMAGE_VERSION:-latest}
-    build: payments
+    build: service/payments
     links:
       - kafka
     ports:
@@ -42,7 +42,7 @@ services:
       SPRING_KAFKA_BOOTSTRAP_SERVERS: kafka:9092
   stock:
     image: wdmk/stock:${IMAGE_VERSION:-latest}
-    build: stock
+    build: service/stock
     links:
       - kafka
     ports:
@@ -51,7 +51,7 @@ services:
       SPRING_KAFKA_BOOTSTRAP_SERVERS: kafka:9092
   users:
     image: wdmk/users:${IMAGE_VERSION:-latest}
-    build: users
+    build: service/users
     links:
       - kafka
     ports:
