+++ service/build.gradle
@@ -1,0 +1,13 @@
+subprojects {
+    apply plugin: "org.springframework.boot"
+    apply plugin: 'io.spring.dependency-management'
+
+    dependencies {
+        implementation 'org.springframework.boot:spring-boot-starter-web'
+        implementation 'org.springframework.kafka:spring-kafka'
+        implementation project(':common')
+        testImplementation 'org.springframework.boot:spring-boot-starter-test'
+        testImplementation 'org.springframework.kafka:spring-kafka-test'
+        testCompile 'org.awaitility:awaitility:3.1.6'
+    }
+}
