+++ service/payments/settings.gradle
