+++ service/stock/Dockerfile
