+++ service/stock/settings.gradle
