+++ service/stock/src/main/java/nl/tudelft/wdm/group1/stock/web/StockController.java
