+++ service/stock/src/main/resources/application.yml
