+++ service/stock/src/test/java/nl/tudelft/wdm/group1/stock/StockApplicationTest.java
