+++ settings.gradle
@@ -6,5 +6,4 @@ pluginManagement {
 
 rootProject.name = 'wdm-kafka-microservices'
 
-include 'users', 'orders', 'stock', 'payments', 'end-to-end', 'common'
-
+include 'service:users', 'service:orders', 'service:stock', 'service:payments', 'end-to-end', 'common'
